refactor: refactor code, transferred files
