+++ src/app/app.js
@@ -1,10 +1,10 @@
 import { swiperСontrol } from './js/slider';
-import { initialize } from './components/product';
-import { sortProduct } from './components/select';
-import { filterProductByCheckbox } from './components/filter';
-import { initFilters } from './components/popupFilter';
-import { initCartPopup } from './components/popupCart';
-import { initializeCart } from './components/productInCart';
+import { initialize } from './js/product';
+import { sortProduct } from './js/select';
+import { filterProductByCheckbox } from './js/filter';
+import { initFilters } from './js/popupFilter';
+import { initCartPopup } from './js/popupCart';
+import { initializeCart } from './js/productInCart';
 
 export default function renderApp() {
   swiperСontrol();
