+++ src/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="https://unpkg.com/swiper/swiper-bundle.min.css" />
+    
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
@@ -67,7 +67,7 @@
      <div class="mobile-title">
       <h2>Краски</h2>
     </div>
-    <div class="swiper-container">
+    <div class="swiper">
       <div class="swiper-wrapper">
         <div class="swiper-slide" >
           
@@ -151,18 +151,18 @@
    <div class="cart-popup">
     <div class="cart-popup__container">
       <div>
-      <div class="cart-popup__title-close">
+      <div class="cart-popup__title-close-container">
       <h3 class="cart-popup__title">Корзина</h3>
       <div class="cart-popup__close-button"></div>
       </div>
-      <div class="cart-popup__quantity-clear">
+      <div class="cart-popup__quantity-clear-container">
         <h3 class="cart-popup__quantity">товара</h3>
         <div class="cart-popup__clear">Очистить список</div>
         </div>
       
          <div class="cart-container"></div>
         </div>
-         <div class="cart-popup__total-order">
+         <div class="cart-popup__total-container">
           <div>
             <span >Итого</span>
             <h2 class="cart-popup__total">12133</h2>
