+++ src/styles/catalog.scss
@@ -1,3 +1,4 @@
+@import 'variables';
 #products-container {
   display: flex;
   flex-wrap: wrap;
@@ -23,12 +24,12 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    font-size: 1.2rem;
+    font-size: $font-size-medium;
     font-weight: 500;
   }
   &__mobile-filter {
     display: none;
-    font-size: 1.2rem;
+    font-size: $font-size-medium;
     font-weight: 500;
     cursor: pointer;
   }
