+++ src/styles/header.scss
@@ -57,7 +57,8 @@
   }
 
   &__tel-link {
-    font-size: 16px;
+    font-size: $font-size-medium;
+    font-weight: 500;
     color: $color-text;
     text-decoration: none;
     transition: color 0.3s;
@@ -67,7 +68,7 @@
     }
   }
   &__tel-link div {
-    font-size: 1rem;
+    font-size: $font-size-small;
     font-weight: 400;
     opacity: 40%;
     color: #1f2020;
