+++ src/styles/product.scss
@@ -24,7 +24,7 @@
   }
 
   &__name {
-    font-size: 1.2rem;
+    font-size: $font-size-medium;
     font-weight: 300;
     color: $color-text;
     margin-bottom: 5px;
@@ -36,10 +36,10 @@
     justify-content: space-between;
     align-items: center;
     width: 100%;
+    font-size: $font-size-medium;
   }
 
   &__price {
-    font-size: 1.2rem;
     font-weight: 600;
     color: $color-text;
   }
@@ -50,7 +50,7 @@
     opacity: 0;
     background-color: #7bb899;
     text-align: center;
-    font-size: 16px;
+
     font-weight: bold;
     border: none;
     border-radius: 8px;
