+++ src/styles/productInCart.scss
@@ -1,4 +1,5 @@
-.cart-container__rectangle {
+@import 'variables';
+.rectangle {
   width: 100%;
   height: 1px;
   background-color: #000000;
@@ -17,7 +18,7 @@
     width: 96px;
   }
   &__name-price {
-    font-size: 1.2rem;
+    font-size: $font-size-medium;
   }
   &__name {
     font-weight: 300;
@@ -33,7 +34,7 @@
     align-items: center;
     gap: 10px;
     font-weight: 400;
-    font-size: 1.2rem;
+    font-size: $font-size-medium;
   }
   &__quantity-counter {
     display: inline-block;
@@ -61,7 +62,7 @@
     height: 40px;
     opacity: 40%;
     border: none;
-    background-color: #ffffff;
+    background-color: $color-background;
     margin-left: 10px;
     cursor: pointer;
   }
