+++ src/styles/select.scss
@@ -1,8 +1,9 @@
+@import 'variables';
 .custom-select {
   position: relative;
   display: inline-block;
   width: 200px;
-  z-index: 20;
+  z-index: 10;
 }
 
 .custom-select__container {
@@ -19,7 +20,7 @@
   border: none;
   font-weight: 500;
   background-color: white;
-  font-size: 1.2rem;
+  font-size: $font-size-medium;
   cursor: pointer;
 
   background-image: url('data:image/svg+xml,%3Csvg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 10 6%22%3E%3Cpath d=%22M5 6L0 0h10z%22/%3E%3C/svg%3E');
@@ -31,32 +32,32 @@
 }
 
 .custom-select__select:hover {
-  background-color: white;
+  background-color: $color-background;
 }
 
 .custom-select__select:focus {
   outline: none;
-  background-color: #7bb899;
+  background-color: $color-accent;
 }
 
 .custom-select--active .custom-select__select {
-  background-color: #7bb899;
+  background-color: $color-accent;
 }
 
 .custom-select__option {
-  font-size: 16px;
+  font-size: $font-size-medium;
 }
 
 .custom-select__select:focus {
   outline: none;
-  border-color: #7bb899;
+  border-color: $color-accent;
 }
 option {
-  background-color: white;
+  background-color: $color-background;
 }
 
 option:hover {
-  background-color: #7bb899;
+  background-color: $color-accent;
 }
 
 .custom-select--asc .custom-select__select {
