+++ src/styles/slider.scss
@@ -1,4 +1,5 @@
-.swiper-container {
+@import 'variables';
+.swiper {
   width: 100%;
   height: 560px;
   position: relative;
@@ -36,12 +37,12 @@ h2 {
 }
 
 p {
-  font-size: 1.2rem;
+  font-size: $font-size-medium;
 }
 
 #prev-slide,
 #next-slide {
-  color: white;
+  color: $color-background;
   padding: 20px;
   border: none;
   border-radius: 5px;
@@ -72,7 +73,7 @@ p {
   color: #1f2020;
 }
 @media (max-width: 768px) {
-  .swiper-container {
+  .swiper {
     display: none;
   }
 }
