+++ src/styles/variables.scss
@@ -1,4 +1,7 @@
 $color-background: #ffffff;
 $color-text: #1f2020;
-
-$font-size-normal: 1rem;
+$color-accent: #7bb899;
+$font-size-small: 1rem;
+$font-size-medium: 1.2rem;
+$font-size-large: 2.1rem;
+$font-size-xlarge: 2.5rem;
